style: Move list quotes out of mako
