+++ components/style/values/specified/mod.rs
@@ -49,6 +49,7 @@ pub use self::length::{LengthOrPercentage, LengthOrPercentageOrAuto};
 pub use self::length::{LengthOrPercentageOrNone, MaxLength, MozLength};
 pub use self::length::{NoCalcLength, ViewportPercentageLength};
 pub use self::length::NonNegativeLengthOrPercentage;
+pub use self::list::Quotes;
 pub use self::outline::OutlineStyle;
 pub use self::rect::LengthOrNumberRect;
 pub use self::percentage::Percentage;
@@ -79,6 +80,7 @@ pub mod gecko;
 pub mod grid;
 pub mod image;
 pub mod length;
+pub mod list;
 pub mod outline;
 pub mod percentage;
 pub mod position;
